Move DocumentRepository to Repository namespace
