+++ lib/Doctrine/ODM/MongoDB/Configuration.php
@@ -15,6 +15,7 @@ use Doctrine\ODM\MongoDB\PersistentCollection\DefaultPersistentCollectionGenerat
 use Doctrine\ODM\MongoDB\PersistentCollection\PersistentCollectionFactory;
 use Doctrine\ODM\MongoDB\PersistentCollection\PersistentCollectionGenerator;
 use Doctrine\ODM\MongoDB\Repository\DefaultRepositoryFactory;
+use Doctrine\ODM\MongoDB\Repository\DocumentRepository;
 use Doctrine\ODM\MongoDB\Repository\RepositoryFactory;
 use function trim;
