+++ tests/Documents/CustomRepository/Repository.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Documents\CustomRepository;
 
-use Doctrine\ODM\MongoDB\DocumentRepository;
+use Doctrine\ODM\MongoDB\Repository\DocumentRepository;
 
 class Repository extends DocumentRepository
 {
