+++ tools/sandbox/Documents/UserRepository.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Documents;
 
-use Doctrine\ODM\MongoDB\DocumentRepository;
+use Doctrine\ODM\MongoDB\Repository\DocumentRepository;
 
 /**
  * UserRepository
